New example showing downloading images
