+++ Makefile
@@ -1,5 +1,6 @@
 SRC_DEPS = src/index.tsx src/third-party/qrcodegen/index.ts
 CFG_DEPS = yarn.lock package.json tsup.config.ts tsconfig.json
+EXAMPLE_DEPS = examples/*.tsx
 
 all: lib/index.js lib/index.d.ts lib/index.js examples/iife/demo.js
 
@@ -15,7 +16,7 @@ lib/esm/index.js: lib $(SRC_DEPS) $(CFG_DEPS)
 lib/index.js: lib $(SRC_DEPS) $(CFG_DEPS)
 	yarn run build:code
 
-examples/iife/demo.js: lib/esm/index.js examples/demo.tsx
+examples/iife/demo.js: lib/esm/index.js ${EXAMPLE_DEPS}
 	yarn run build:examples
 
 clean:
